// standalone-extension/src/state/types.ts
export interface Profile {
  id: string;
  name: string;
  keywordsInclude: string[];
  keywordsExclude: string[];
  experience: string[];
  schedule: string[];
  employment: string[];
  regions?: string[];
  salary?: {
    amount?: number;
    currency?: string;
  };
  coverLetterTemplate?: string;
  selectedResumeHash?: string | null;
  createdAt: number;
  updatedAt: number;
}